refactor: reduce the max allowed chunkSize
This is important now since we store the same number of items in the chunk command
 in the RocksDB state chunk block.

RocksDB is configured with a block size of 32KB,
so chunks significantly larger than this would span multiple blocks,
which is not optimal for read/write performance.

A chunk size of 3000 item keys results in a chunk size of approximately 24KB (31KB with overhead),
which is reasonable for RocksDB to handle efficiently.

Larger chunk sizes could lead to increased memory usage
and potential performance degradation,
while smaller chunk sizes would increase the number of database entries,
leading to higher overhead in managing them.

(cherry picked from commit 53190ca78301f668737ea8305bbca2834c7193ac)